Sistema de Login p4 - Protejendo rotas, usuário deslogado

+++ App/views/home/index.php
@@ -13,10 +13,15 @@ if(!empty($data['mensagem'])){
 
 <?php foreach ($data['registros'] as $note){ ?>
 
-    <h1> <a href="/notes/ver/<?php echo $note['ID']; ?>"> <?php echo $note['ID'].' - '.$note['TITULO']; ?> </h1><p>
-    </a> <a href="/notes/editar/<?php echo $note['ID']; ?>">Editar</a> |
-    </a> <a href="/notes/excluir/<?php echo $note['ID']; ?>">Excluir</a> 
+    <h1> <a href="/notes/ver/<?php echo $note['ID']; ?>"> <?php echo $note['ID'].' - '.$note['TITULO']; ?> </h1><p></a>
+    
+    <?php if(isset($_SESSION['logado'])): ?> <!-- Se estiver logado habilita "Editar e Excluir" -->
 
+
+        <a href="/notes/editar/<?php echo $note['ID']; ?>">Editar</a> |
+        </a> <a href="/notes/excluir/<?php echo $note['ID']; ?>">Excluir</a> 
+
+   <?php endif; ?>
